Add date/regex tests.
