+++ src/NEW/index.ts
@@ -7,15 +7,17 @@ import * as criteria from './criteria';
 import createSchema from './createSchema';
 import { array } from './schemas/array';
 import { bool } from './schemas/bool';
-import { instance, date, regex } from './schemas/instance';
+import { date } from './schemas/date';
+import { instance } from './schemas/instance';
 import { number } from './schemas/number';
 import { object } from './schemas/object';
+import { regex } from './schemas/regex';
 import { schema, blueprint } from './schemas/schema';
 import { string } from './schemas/string';
 
 export { criteria, createSchema };
 
-export const predicates = {
+export default {
   array,
   blueprint,
   bool,
@@ -35,8 +37,10 @@ export const predicates = {
 
 export * from './schemas/array';
 export * from './schemas/bool';
+export * from './schemas/date';
 export * from './schemas/instance';
 export * from './schemas/number';
+export * from './schemas/regex';
 export * from './schemas/schema';
 export * from './schemas/string';
 export * from './schemas/object';
