+++ src/NEW/schemas/date.ts
@@ -1,0 +1,5 @@
+import { instance } from './instance';
+
+export function date() /* infer */ {
+  return instance().of(Date);
+}
