+++ src/NEW/schemas/regex.ts
@@ -1,0 +1,5 @@
+import { instance } from './instance';
+
+export function regex() /* infer */ {
+  return instance().of(RegExp);
+}
